container/requirements.txt
# PyPI library used in all project
rdetoolkit==1.3.4

# PyPI libraries to be customized and installed in each project
# python-magic==0.4.25 ...etc
scipy==1.16.1
pyrtools==1.0.9
pandas-stubs==2.3.2.250827
scipy-stubs==1.16.2.0